requirements.txt
fastapi
uvicorn
torch
Pillow
python-multipart
ultralytics